RELEASE
DISTRIBUTED SOURCE CODING USING SYNDROMES.
@author: BLANCO CAAMANO, RAMON.
@about:
In this report, we consider the related problem of compressing a source which is correlated with another source that is available only at the decoder. This problem has been studied in the information theory literature under the name of the Slepian{Wolf source coding problem for the lossless coding case, and as "rate-distortion with side-information" for the lossy coding case.
In this work, a practical implementation in Matlab is provided based on algebraic trellis codes dubbed as Distributed Source Coding Using Syndromes (DISCUS), that can be applicable in a variety of settings. Construction detail are shown in paper Distributed Source Coding Using Syndromes (DISCUS): Design and Construction (see References), in particular, the section III.C.3(using the trellis described in Figure 4) that is what concerns us in this report.
Finally, main results are shown concluding with the study and comparison of the following proposed scenarios:
 *Case 1: Coding and decoding of X when Y is known at reception.
 *Case 2: Coding and decoding of X when Y is independent.
 *Case 3: Coding and decoding of X when Y is known at both sides.
 *Case 4: Estimation of X exclusively based on Y.